Moved the DOIOstiWebParser classes into osti.py and removed the osti_web_parser.py module
